requirements.txt
matplotlib==3.4.2
numpy==1.18.2
pandas==1.3.1
Pillow==8.4.0
scikit_learn==1.0.1
scipy==1.7.1
torch==1.10.0
torchvision==0.11.12
